[gap-decorations] Interpolation with repeaters for *-rule-width
This CL implements Interpolation for *-rule-width for repeaters.

In order to handle repeaters, we add a new InterpolableType
`InterpolableGapDataRepeater` which handles interpolation for repeaters
by essentially keeping a member InterpolableList of the values in the
repeater (for now just `InterpolableLength`) and interpolating those
individually through that type.

The overall structure of interpolation for gap decorations will be as
follows:

`CSSGapLengthListInterpolationType` handles interpolation for
`GapDataList` We then hand off the actual interpolation with an
`InterpolableList` of `InterpolableLength`,
`InterpolableGapDataRepeater` (which itself contains an
`InterpolableList` of `InterpolableLengths`) and `InterpolableColor`.

InterpolableGapDataRepeater is templated, in order to handle for
interpolation for column/row-rule-color. This will be implemented in a
follow-up CL.

Bug: 357648037, 419066541

Change-Id: Ia43396607d0bbe3a6266571df31ae36feb882875