[CompositeClipPathAnimation] Add test for calc() with unit update
To fix a gap in coverage where changes in width/height and calc() were
not tested for composited clip path animations. Although believed to be
settled now, mixed units were at one time a pain point for the feature,
before the transition to SKIA paths rather than storing interpolated
values. This test serves as an extra guardrail for regression.

Bug: 40503852
Change-Id: I15dac37f9ddd65b4aaac73c1424ad7aa154ce28f